Update class.core.js
I am not sure if there are any side effects here but this allows classes that extend other classes to maintain there constructors. This is helpful for looking up the classname, something like myInstance.__proto__.constructor.name. If you write over the constructor with an anonymous function we lose that lookup. 

This might be a horrible idea but just throwing it out there :)